+++ test/queue_executor_test.dart
@@ -11,6 +11,7 @@ void main() {
 
   tearDown(() {
     _delayValues = [];
+    executor.dispose();
   });
 
   group("executes tasks:", () {